+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/tests/bulkloading/AwsSecretsManagerPerformanceAcceptanceTest.java
@@ -19,8 +19,8 @@ import tech.pegasys.web3signer.AwsSecretsManagerUtil;
 import tech.pegasys.web3signer.common.config.AwsAuthenticationMode;
 import tech.pegasys.web3signer.dsl.signer.SignerConfigurationBuilder;
 import tech.pegasys.web3signer.signing.KeyType;
-import tech.pegasys.web3signer.signing.config.AwsSecretsManagerParameters;
-import tech.pegasys.web3signer.signing.config.AwsSecretsManagerParametersBuilder;
+import tech.pegasys.web3signer.signing.config.AwsVaultParameters;
+import tech.pegasys.web3signer.signing.config.AwsVaultParametersBuilder;
 import tech.pegasys.web3signer.tests.AcceptanceTestBase;
 
 import java.net.URI;
@@ -120,8 +120,8 @@ public class AwsSecretsManagerPerformanceAcceptanceTest extends AcceptanceTestBa
 
   @Test
   void largeNumberOfKeysAreLoadedSuccessfully() {
-    final AwsSecretsManagerParameters awsSecretsManagerParameters =
-        AwsSecretsManagerParametersBuilder.anAwsSecretsManagerParameters()
+    final AwsVaultParameters awsVaultParameters =
+        AwsVaultParametersBuilder.anAwsParameters()
             .withAuthenticationMode(AwsAuthenticationMode.SPECIFIED)
             .withRegion(AWS_REGION)
             .withAccessKeyId(RO_AWS_ACCESS_KEY_ID)
@@ -133,7 +133,7 @@ public class AwsSecretsManagerPerformanceAcceptanceTest extends AcceptanceTestBa
     final SignerConfigurationBuilder configBuilder =
         new SignerConfigurationBuilder()
             .withMode("eth2")
-            .withAwsSecretsManagerParameters(awsSecretsManagerParameters)
+            .withAwsParameters(awsVaultParameters)
             .withStartupTimeout(STARTUP_TIMEOUT)
             .withLogLevel(Level.INFO);