+++ core/src/main/java/tech/pegasys/web3signer/core/config/Eth1Config.java
@@ -14,6 +14,7 @@ package tech.pegasys.web3signer.core.config;
 
 import tech.pegasys.web3signer.core.config.client.ClientTlsOptions;
 import tech.pegasys.web3signer.core.service.jsonrpc.handlers.signing.ChainIdProvider;
+import tech.pegasys.web3signer.signing.config.AwsVaultParameters;
 import tech.pegasys.web3signer.signing.config.AzureKeyVaultParameters;
 import tech.pegasys.web3signer.signing.config.KeystoresParameters;
 
@@ -43,6 +44,8 @@ public interface Eth1Config {
   ChainIdProvider getChainId();
 
   AzureKeyVaultParameters getAzureKeyVaultConfig();
+
+  AwsVaultParameters getAwsVaultParameters();
 
   long getAwsKmsClientCacheSize();