+++ signing/src/main/java/tech/pegasys/web3signer/signing/bulkloading/BlsAwsBulkLoader.java
@@ -20,15 +20,15 @@ import tech.pegasys.web3signer.keystorage.common.MappedResults;
 import tech.pegasys.web3signer.signing.ArtifactSigner;
 import tech.pegasys.web3signer.signing.BlsArtifactSigner;
 import tech.pegasys.web3signer.signing.config.AwsSecretsManagerFactory;
-import tech.pegasys.web3signer.signing.config.AwsSecretsManagerParameters;
+import tech.pegasys.web3signer.signing.config.AwsVaultParameters;
 import tech.pegasys.web3signer.signing.config.metadata.SignerOrigin;
 
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 
-public class AWSBulkLoadingArtifactSignerProvider {
+public class BlsAwsBulkLoader {
 
-  public MappedResults<ArtifactSigner> load(final AwsSecretsManagerParameters parameters) {
+  public MappedResults<ArtifactSigner> load(final AwsVaultParameters parameters) {
     try (final AwsSecretsManagerProvider awsSecretsManagerProvider =
         new AwsSecretsManagerProvider(parameters.getCacheMaximumSize())) {
       final AwsSecretsManager awsSecretsManager =